Remove Bromite's autoupdater
Chromium isn't given the permissions to install apps on the panel. So
the autoupdating process will just loop saying "an update is
available" then fail to update it.

In any case, the update comes from Bromite and not Promethean. While
autoupdate may be nice it's probably not a huge advantage to us and
carries the risk of bringing in new bugs that'll cause failures.

Done as part of AP9-1561