Move local patches higher in the patchlist
This is to allow for a clean rebase of the latest upstream changes
onto the existing local master.

The local force-tablet-mode patch depends on the existing
force-tablet-UI patch so that's also moved up. The force-desktop-mode
patch depends on the user-agent-customization patch so that's put in
the right place as well.

Done as part of AP9-2571

Move desktop mode patch down a bit

as it depends on User-agent-customization.patch