AP9-2791: Adjust the text scaling value of Chromium to the maximum value, and the chrome homepage page is stretched after restarting the panel.
When the value of Text scaling is large enough and the checkbox of Force
enable zoom is checked, then if we restart the panel or kill Chromium
and reopen Chromium, the web page will be stretched. The effect is like
that stretch the page with two fingers.
This commit ensures that when the checkbos of Force enable zoom is checked, the
web page will not be stretched.