AP9-10932 Development - add an API of deleting files in Chromium
Fix version: 1.2.1
Per https://prometheanworld.atlassian.net/browse/AP9-10282?focusedCommentId=252420, if we show the 'Settings' at the bottom of the menu after signing in to Promethean, we need to delete some other files or dirs to ensure that the user data is not leaked. If the deleted file list is defined in Chromium, and if we want to delete other files, we need to modify the Chromium code, this will involve the work of two teams (IDP and AP9), so I think Chromium can provide the API of deleting files or dirs, IDP decides which files or dirs to delete, and it will be easier for IDP to test.