AP9-10449 AP9 - Camera not detected in Chromium browser for Google Meet
Fix version = 1.2.1

Why:
There is a big difference between the USB camera connected in the user scenario of the ActivPanel and the embedded camera in the mobile phone.
For some outdated or low-tier USB cameras, valid parameters cannot be obtained through the CameraCharacteristics object, including REQUEST_AVAILABLE_CAPABILITIES, SCALER_STREAM_CONFIGURATION_MAP, and so on.
If we still try to invoke the NULL object after getting it through CameraCharacteristics, the Null-pointer exception will be produced, causing Chromium to crash.

How:
In VideoCaptureCamera2, add a non-NULL judgment before getting related properties through CameraCharacteristics and trying to invoke them.