AP9-10977 Chrome's homepage does not launch with selected in homepage settings
Fix version: 1.2.1
why:
The 'Set-default-page-url' and 'Modify-default-url-AP9-5158' patches change the default home url to
'https://www.prometheanworld.com/', and we found if 'Settings->Homepage' is toggled off and
'Settings->Homepage->Chrome's homepage' is selected, they don't seem to work, it is caused by
changing the 'UrlConstants->NTP_URL' and 'UrlConstants->NTP_NON_NATIVE_URL' values.
How:
1. Revert the 'UrlConstants->NTP_URL' and 'UrlConstants->NTP_NON_NATIVE_URL' values.
2. Set the default value of 'ChromePreferenceKeys.HOMEPAGE_PARTNER_CUSTOMIZED_DEFAULT_URI' to
'https://www.prometheanworld.com/'.
3. Set the default 'Use for new tabs' toggle to on.
4. Fix the issue that 'New Incognito tab' doesn't open the home page.
Relate to the AP9-10977 and AP9-10970 tickets.
Relate to the 'Add-option-to-use-home-page-as-NTP.patch' patch.