AP9-11548 Drag and Drop function in Chromium Browser does not work
Fix version: 1.4.0
'Enable-StrictOriginIsolation-and-SitePerProcess.patch' and 'Enable-SPPI-for-devices-with-enough-memory.patch' will cause the issue that Drag and Drop funtion doesn't work ,so revert them.
For the Chromium that the version is 100+ and OS is Android, the Drag and Drop function can transfer text, link and image, but we can just get one of them, and the priority is that image > link > text, so if we set the text, link and image data, we can just get the image data, and we can't get the text and image data, but on 99-, the Drag and Drop function only supports transferring text, so the web developers will be confused for that. In order to be compatible with the old code, we will disable transferring link and image.